Initial refactor of the main mod class code. - Break out major portions of logic into helper functions - Pre-configure data in rough db-like way to reduce code duplication
