+++ SummonersAssociation.csproj
@@ -14,7 +14,7 @@
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <PlatformTarget>AnyCPU</PlatformTarget>
+    <PlatformTarget>x86</PlatformTarget>
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
@@ -62,11 +62,11 @@
       <HintPath>..\..\XNA Framework Dlls\Microsoft.Xna.Framework.Xact.dll</HintPath>
     </Reference>
     <Reference Include="ReLogic">
-      <HintPath>..\..\..\Modding\tModLoader\references\ReLogic.dll</HintPath>
+      <HintPath>..\..\tModLoader\references\ReLogic.dll</HintPath>
     </Reference>
     <Reference Include="System" />
-    <Reference Include="tModLoaderDebug">
-      <HintPath>C:\Program Files (x86)\Steam\steamapps\common\Terraria\tModLoaderDebug.exe</HintPath>
+    <Reference Include="Terraria">
+      <HintPath>..\..\tModLoader\tModLoader.Windows.v0.10.1.1\Terraria.exe</HintPath>
     </Reference>
   </ItemGroup>
   <ItemGroup>
