+++ SummonersAssociation.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2010
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SummonersAssociation", "SummonersAssociation.csproj", "{37E19856-3C3D-478E-A2AA-2A22FF516E95}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{37E19856-3C3D-478E-A2AA-2A22FF516E95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{37E19856-3C3D-478E-A2AA-2A22FF516E95}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{37E19856-3C3D-478E-A2AA-2A22FF516E95}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{37E19856-3C3D-478E-A2AA-2A22FF516E95}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8C18A2B0-9AB3-474D-9DE2-D470E533C261}
+	EndGlobalSection
+EndGlobal
