+++ build.txt
@@ -1,5 +1,5 @@
 author = jopojelly
-version = 0.2.4
+version = 0.3.0
 displayName = Summoners Association
 homepage = http://forums.terraria.org/index.php?threads/summoners-association.30041/
 buildIgnore = .vs\*, *.csproj, *.user, obj\*, bin\*, *.config, .git\*, README.md, .gitignore
