+++ description.txt
@@ -1,4 +1,4 @@
-﻿This mod aims to fix a few grievances with summons and flesh out Summoner to be a fully developed class. This vision of Summoner is a summoner who has his hands full buffing and controlling his minions and has little time for weapons himself.
+﻿This mod aims to fix a few grievances with summons and flesh out Summoner to be a fully developed class. This vision is of a summoner who has their hands full buffing and controlling minions and has little time for other weapons.
 
 Includes 3 items:
